# front-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OOUN Restaurant API Documentation",
        "description": "University Restaurant Management System API - OOUN Tunisia",
        "contact": {
            "name": "OOUN Restaurant System",
            "email": "admin@ooun.tn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/admin/menus": {
            "get": {
                "tags": [
                    "Admin - Menu Management"
                ],
                "summary": "Get all menus",
                "description": "Retrieve paginated list of all weekly menus with optional filters",
                "operationId": "36051cda615bfd27a42023c1d63db66b",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "unavailable"
                            ]
                        }
                    },
                    {
                        "name": "day",
                        "in": "query",
                        "description": "Filter by day",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Mon",
                                "Tue",
                                "Wed",
                                "Thu",
                                "Fri",
                                "Sat",
                                "Sun"
                            ]
                        }
                    },
                    {
                        "name": "meal_type",
                        "in": "query",
                        "description": "Filter by meal type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "breakfast",
                                "lunch",
                                "dinner"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menus retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Menu Management"
                ],
                "summary": "Create new menu",
                "description": "Create a new weekly menu item",
                "operationId": "d9721b079957d3a5b13d96309a833cd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "day",
                                    "meal_type",
                                    "title"
                                ],
                                "properties": {
                                    "day": {
                                        "type": "string",
                                        "enum": [
                                            "Mon",
                                            "Tue",
                                            "Wed",
                                            "Thu",
                                            "Fri",
                                            "Sat",
                                            "Sun"
                                        ],
                                        "example": "Mon"
                                    },
                                    "meal_type": {
                                        "type": "string",
                                        "enum": [
                                            "breakfast",
                                            "lunch",
                                            "dinner"
                                        ],
                                        "example": "lunch"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Couscous"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Traditional couscous with vegetables"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "vegetarian",
                                            "halal"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "unavailable"
                                        ],
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/menus/{id}": {
            "get": {
                "tags": [
                    "Admin - Menu Management"
                ],
                "summary": "Get specific menu",
                "description": "Retrieve detailed information about a specific menu including feedback and reservations",
                "operationId": "3e117a125b402ab5a105789059850437",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Menu Management"
                ],
                "summary": "Update menu",
                "description": "Update an existing menu item",
                "operationId": "1bcdaefc8a35dfce0f92736a82f865a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "day": {
                                        "type": "string",
                                        "enum": [
                                            "Mon",
                                            "Tue",
                                            "Wed",
                                            "Thu",
                                            "Fri",
                                            "Sat",
                                            "Sun"
                                        ]
                                    },
                                    "meal_type": {
                                        "type": "string",
                                        "enum": [
                                            "breakfast",
                                            "lunch",
                                            "dinner"
                                        ]
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "unavailable"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Menu Management"
                ],
                "summary": "Delete menu",
                "description": "Remove a menu item from the system",
                "operationId": "272f57425e4df4f737b526e7ebd5aa38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Menu not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}